Organize components into folder
