+++ app/(availability)/EditAvailability.js
@@ -10,16 +10,16 @@ import {
 	TextInput,
 } from 'react-native';
 
-import { theme } from '../styles/theme';
+import { theme } from '../../styles/theme';
 
-import { UserContext } from '../context/UserContext';
-import DaySchedule from '../components/availability/DaySchedule';
-import TimePicker from '../components/availability/TimePicker';
-import CustomModal from '../components/common/CustomModal';
-import Button from '../components/common/Button';
-import MeetingConfig from '../components/availability/MeetingConfig';
-import CustomPicker from '../components/availability/CustomPicker';
-import PricingConfig from '../components/availability/PricingConfig';
+import { UserContext } from '../../context/UserContext';
+import DaySchedule from '../../components/availability/DaySchedule';
+import TimePicker from '../../components/availability/TimePicker';
+import CustomModal from '../../components/common/CustomModal';
+import Button from '../../components/common/Button';
+import MeetingConfig from '../../components/availability/MeetingConfig';
+import CustomPicker from '../../components/availability/CustomPicker';
+import PricingConfig from '../../components/availability/PricingConfig';
 
 const BREAK_TIME = 'BREAK_TIME';
 const START_TIME_INCREMENT = 'START_TIME_INCREMENT';
