+++ app/(meeting-dashboard)/MeetingDashboard.js
@@ -1,10 +1,10 @@
 import React, { useContext } from 'react';
 import { Stack, router } from 'expo-router';
 import { View, Text, StyleSheet, SafeAreaView } from 'react-native';
-import { UserContext } from '../context/UserContext';
-import Button from '../components/common/Button';
-import { theme } from '../styles/theme';
-import { convertMinutesToTime } from '../utils/dateAndTime';
+import { UserContext } from '../../context/UserContext';
+import Button from '../../components/common/Button';
+import { theme } from '../../styles/theme';
+import { convertMinutesToTime } from '../../utils/dateAndTime';
 
 const MeetingDashboard = () => {
 	const { userData } = useContext(UserContext);
